Modularize 'util.hpp'
